+++ app/data/repositories/__init__.py
@@ -1,9 +1,9 @@
-from app.data.repositories.common_repository import CommonRepository
+from app.data.repositories.common import CommonRepository
 from app.data.repositories.layer0 import INTERNAL_ID_COLUMN_NAME, Layer0Repository
-from app.data.repositories.layer1_repository import Layer1Repository
+from app.data.repositories.layer1 import Layer1Repository
 from app.data.repositories.layer2 import Layer2Repository
-from app.data.repositories.queue_repository import QueueRepository
-from app.data.repositories.tmp_data_repository_impl import TmpDataRepositoryImpl
+from app.data.repositories.queue import QueueRepository
+from app.data.repositories.tmp_data import TmpDataRepositoryImpl
 
 __all__ = [
     "CommonRepository",
